refactor some telescope config into telescope/utils.lua
